location fixed qibla pas encore
git-svn-id: svn+ssh://jabouzi.com/home/jabouzic/svn/qsalat/trunk@19 ea0bdeb7-57c4-4e52-a4b7-181e3d1c9d5a

+++ src/qlocation.h
@@ -48,7 +48,7 @@ public:
     
 private slots:
     void loadAddress(QString);
-    void loadCoordinates(float,float);
+    void loadCoordinates(double,double);
     void showItem();
     void updateLatLng();
     void apply();
@@ -67,8 +67,8 @@ private:
     QList<QPointF> coordinates;
     DomParser parser;
     QString file;    
-    float latitude;
-    float longitude;
+    double latitude;
+    double longitude;
     QString country;
     QString city;
     int timezone;
